requirements.txt
flask
flask-cors
requests
transformers
torch
nltk
python-docx
numpy
python-dotenv